created and updated column added to database

+++ server.js
@@ -23,7 +23,7 @@ app.get("/notes",async (req, res)=>{
 
 app.get("/setup",async (req, res)=>{
     try{
-        await pool.query('CREATE TABLE notes(id SERIAL PRIMARY KEY, title VARCHAR(200), content VARCHAR(1000))')
+        await pool.query('CREATE TABLE notes(id SERIAL PRIMARY KEY, title VARCHAR(200), content VARCHAR(1000), createdAt VARCHAR(30), updatedAt VARCHAR(30))')
         res.status(200).send({ message: "Successfully created table" })
     }catch(e){
         console.log(e)
@@ -33,6 +33,9 @@ app.get("/setup",async (req, res)=>{
 
 app.post("/delete",async (req, res)=>{
     var {id} = req.body
+    if(!id){
+        return res.status(400).send({message: "missing required field id"})
+    }
     try{
         await pool.query(`DELETE FROM notes WHERE id = ${id}`)
         res.status(200).send({ message: "note deleted successfully" })
@@ -50,8 +53,9 @@ app.post("/update",async (req, res)=>{
     }
 
     try{
-        var query = "UPDATE notes SET title=$1, content=$2 WHERE id=$3"
-        var values = [title, content, id]
+        var updatedAt = new Date()
+        var query = "UPDATE notes SET title=$1, content=$2, updatedAt=$4 WHERE id=$3"
+        var values = [title, content, id, updatedAt]
         
         await pool.query(query, values)
         // await pool.query(`UPDATE notes SET title=${title}, content=${content} WHERE id = ${id}`)
@@ -64,8 +68,12 @@ app.post("/update",async (req, res)=>{
 
 app.post('/insert', async (req, res) => {
     const { title, content } = req.body
+    if(!title || !content){
+        return res.status(400).send({message: "missing required fields"})
+    }
     try {
-        await pool.query('INSERT INTO notes (title, content) VALUES ($1, $2)', [title, content])
+        var createddAt = new Date()
+        await pool.query('INSERT INTO notes (title, content, createdAt, updatedAt) VALUES ($1, $2, $3, $4)', [title, content,createddAt,createddAt])
         res.status(200).send({ message: "Successfully added note" })
     } catch (err) {
         console.log(err)
